requirements.txt
torch
transformers 
open3d 
pillow 
numpy 
matplotlib
torchvision